requirements.txt
googletrans
pytube
streamlit
youtube_dl
lyricsgenius
google-api-python-client
wikipedia-api
groq
python-dotenv
wikipedia
isodate